ValidIntegers: move to the new `Numbers` category
This is a BC-break and should be annotated as such in the changelog.
