+++ PHPCompatibility/Tests/Numbers/ValidIntegersUnitTest.inc
